requirements.txt
flask==2.3.3
flask-restx==1.2.0
Werkzeug==2.3.8
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv
pytest-cov==4.1.0
pytest-xdist==3.5.0